Minor changes. Somewhat stable. Severely lacking documentation.
Began writing custom filters for the index page such that it will be
no longer necessary to use the widthratio template tag for computing
uptime in days and bandwidth in KB/s, but there are type issues with
this that I can't figure out -- similar filters written for the details
page work perfectly. Also began figuring out how to enable custom
searches, but decided to hold off on this until raw SQL queries are
removed and queries that use Django's ORM are implemented.