Many minor fixes/improvements. See below.
Gzip compression is now used for HTTP responses, reducing the size of
the index page from ~3.2MB to ~340KB. A baseindex.html page was added
such that CSV links can only appear on index-related pages. Many
try/except blocks were added where it is necessary to use a foreignkey
relationship. I am not sure if there is a way to do this without a
try/except block.

The caching schema may or may not be working. I will have to wait until
descriptors are imported from a local Tor data directory to test
further.

.gifs were converted into .pngs. I expected a more significant decrease
in size.