Fixed critical cache.sql bug, see below.
Prior to this commit, triggers were not written correctly: data added to
tordir.cache was not added to tordir.public. This was due to a 'BEFORE
INSERT OR UPDATE' trigger not returning the row used in the trigger. I
opted to make these triggers AFTER triggers to minimize TorStatus'
impact on Tor Metrics; if it is decided that it is more important that
TorStatus get the statusentry and descriptor data first, this can be
changed, but I should be careful to return the NEW row in this case.