Lots of code cleanup/documentation; IP sort as IP.
settings.template was cleaned up to be more intuitive.

cache.sql was modified to cast an IP address to an INET column, so now
the IP address column in the index page can be sorted as an IP address
rather than a string in lexicographic order. e.g. 255.255.255.255 is
higher than 99.99.99.99 instead of the other way around.

The logic in many templates was cleaned up.

status/tmp was added with the help of a .gitignore file.